+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# 2019-09-05
+
+- Added the conversion of quakeml xml data to a pandas dataframe.
+
 # 2019-09-04
 
-This is the initial version.
+- This is the initial version.